requirements.txt
python-telegram-bot==22.6
openai
python-dotenv